experimental/LISA-diffusion/requirement.txt
accelerate>=0.16.0
torchvision
transformers>=4.25.1
ftfy
peft
wandb
tensorboard
Jinja2
webdataset
datasets